online_shop/requirements.txt
Django==2.2.10
django-ckeditor==5.6.1
